Fix discrepency in RichPayloadEventSource.Keywords by moving them to a shared file.
