+++ src/Core/Managed/Net46/Extensibility/Implementation/RichPayloadEventSource.cs
@@ -19,7 +19,7 @@
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing;
 #endif
-    internal sealed class RichPayloadEventSource : IDisposable
+    internal sealed partial class RichPayloadEventSource : IDisposable
     {
         /// <summary>RichPayloadEventSource instance.</summary>
         public static readonly RichPayloadEventSource Log = new RichPayloadEventSource();
@@ -290,57 +290,6 @@
                     new EventSourceOptions { ActivityOptions = EventActivityOptions.Recursive, Keywords = Keywords.Operations, Opcode = eventOpCode, Level = EventLevel.Informational },
                     payload);
             }
-        }
-
-        /// <summary>
-        /// Keywords for the RichPayloadEventSource.
-        /// </summary>
-        public sealed class Keywords
-        {
-            /// <summary>
-            /// Keyword for requests.
-            /// </summary>
-            public const EventKeywords Requests = (EventKeywords)0x1;
-
-            /// <summary>
-            /// Keyword for traces.
-            /// </summary>
-            public const EventKeywords Traces = (EventKeywords)0x2;
-
-            /// <summary>
-            /// Keyword for events.
-            /// </summary>
-            public const EventKeywords Events = (EventKeywords)0x4;
-
-            /// <summary>
-            /// Keyword for exceptions.
-            /// </summary>
-            public const EventKeywords Exceptions = (EventKeywords)0x8;
-
-            /// <summary>
-            /// Keyword for dependencies.
-            /// </summary>
-            public const EventKeywords Dependencies = (EventKeywords)0x10;
-
-            /// <summary>
-            /// Keyword for metrics.
-            /// </summary>
-            public const EventKeywords Metrics = (EventKeywords)0x20;
-
-            /// <summary>
-            /// Keyword for page views.
-            /// </summary>
-            public const EventKeywords PageViews = (EventKeywords)0x40;
-
-            /// <summary>
-            /// Keyword for operations (Start/Stop).
-            /// </summary>
-            public const EventKeywords Operations = (EventKeywords)0x80;
-
-            /// <summary>
-            /// Keyword for availability.
-            /// </summary>
-            public const EventKeywords Availability = (EventKeywords)0x200;
         }
     }
 }
